FortniteESPHack.cpp
Fortnite ESP Hack
This hack allows the player to see the positions of all enemies on the map. To achieve this, the game's memory is scanned for enemy player data, and the positions are displayed on the screen. Note that this hack may result in your account being banned from the game.